requirements.txt
FlagEmbedding
accelerate
asyncio
beautifulsoup4
chromadb
crawl4ai
devtools
docling
fastapi
gpt4all
gradio
httpx
huggingface_hub
lancedb
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-openai
langchain_community
langchain_core
langchain_milvus
langchain_openai
langchainhub
langgraph
langserve
logfire
ollama
openai
pydantic
pymilvus
python-dotenv
requests
sentence-transformers
sse_starlette
streamlit
supabase
tavily-python
tiktoken
torch
transformers
typing_extensions
uvicorn
watchdog
yfinance
scrapy
qdrant_client
